break out common api code
